Use `&self` for handler trait
