+++ src/handler.rs
@@ -3,16 +3,16 @@ use crate::response::IntoResponse;
 use crate::{RequestContext, Response};
 
 pub trait Handler<Arg = (), Ret = ()> {
-    fn handle(this: Self, req: RequestContext) -> Response;
+    fn handle(&self, req: RequestContext) -> Response;
 }
 
 impl<F, R> Handler<(), R> for F
 where
-    F: FnOnce() -> R,
+    F: Fn() -> R,
     R: IntoResponse,
 {
-    fn handle(this: Self, _req: RequestContext) -> Response {
-        this().into_response()
+    fn handle(&self, _req: RequestContext) -> Response {
+        self().into_response()
     }
 }
 
@@ -21,14 +21,14 @@ macro_rules! impl_handler {
         #[allow(unused_parens)]
         impl<F, $arg1, $( $( $args, )*)? R> Handler<($arg1$(, $( $args, )*)?), R> for F
         where
-            F: FnOnce($arg1$(, $( $args, )*)?) -> R,
+            F: Fn($arg1$(, $( $args, )*)?) -> R,
             $arg1: FromOwnedRequest,
             $( $( $args: FromRequest, )* )?
             R: IntoResponse,
         {
 
             #[allow(unused_mut, unused_variables)]
-            fn handle(this: Self, mut req: RequestContext) -> Response {
+            fn handle(&self, mut req: RequestContext) -> Response {
                 paste::paste! {
                     $($(
                         let [< $args:lower >] = match <$args as FromRequest>::from_request(&mut req) {
@@ -40,7 +40,7 @@ macro_rules! impl_handler {
                         Ok(e) => e,
                         Err(err) => return err.into_response(),
                     };
-                    this(e1 $(, $( [< $args:lower >] ),*)?).into_response()
+                    self(e1 $(, $( [< $args:lower >] ),*)?).into_response()
                 }
             }
         }
