+++ src/lib.rs
@@ -7,27 +7,29 @@ use lunatic::{Mailbox, Process};
 pub use submillisecond_macros::*;
 
 pub use crate::error::*;
+pub use crate::guard::*;
 pub use crate::handler::*;
 pub use crate::request::*;
-pub use crate::response::*;
+use crate::response::{IntoResponse, Response};
 
 #[macro_use]
 pub(crate) mod macros;
 
-pub mod core;
+mod core;
 pub mod defaults;
 pub mod extract;
-pub mod guard;
 #[cfg(feature = "json")]
 pub mod json;
 pub mod params;
+pub mod reader;
+pub mod response;
 #[cfg(feature = "template")]
 pub mod template;
 
 mod error;
+mod guard;
 mod handler;
 mod request;
-mod response;
 
 /// Signature of router function generated by the [`router!`] macro.
 pub type Router = fn(RequestContext) -> Response;
@@ -66,7 +68,7 @@ impl Application {
                     let http_version = request.version();
 
                     let mut response =
-                        Handler::handle(handler, RequestContext::from(request)).into_response();
+                        Handler::handle(&handler, RequestContext::from(request)).into_response();
 
                     let content_length = response.body().len();
                     *response.version_mut() = http_version;
