+++ src/request.rs
@@ -1,7 +1,7 @@
 use std::{convert, ops};
 
-use crate::core::UriReader;
 use crate::params::Params;
+use crate::reader::UriReader;
 use crate::Response;
 
 /// Wrapper for [`http::Request`] containing params and cursor.
